requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
frozenlist==1.3.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
Jinja2==3.1.2
kaitaistruct==0.9
lxml==4.8.0
MarkupSafe==2.1.1
multidict==6.0.2
outcome==1.1.0
packaging==21.3
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-geoip==1.2
PyYAML==6.0
redis==4.3.1
requests==2.27.1
selenium==4.1.5
selenium-wire==4.6.4
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
Werkzeug==2.0.3
wrapt==1.14.1
wsproto==1.1.0
yarl==1.7.2
zstandard==0.17.0
